Added pose description pop up and Icon buttons
Switched to internal storage for the poeses to be able to edit them
Also added a pop up if you tap on a pose in Exercises (May need to name them the same)
In the pop up there is the possibility to edit the pose. Currently only the texts. Image needs to be done
Groups are not yet visible in the pop up